+++ Rag_5040/SAM.lua
@@ -60,7 +60,6 @@ local sets = {
         Hands = 'Gavial Fng.Gnt. +1',
         Ring1 = 'Shadow Ring',
         Ring2 = 'Merman\'s Ring',
-        Back = 'Resentment Cape',
         Legs = 'Coral Cuisses +1',
         Feet = 'Coral Greaves +1',
     },