+++ common/changelog.md
@@ -591,3 +591,4 @@
 - NIN - Allow TP_Aftermath to re-apply to work with how weapon loadouts are intended to be used
 - MNK - Add a focus set modifier for hundred fists used if in LowAcc mode
 - WAR - Add Jump for /DRG
+- All Jobs - Moves resentment cape to gcinclude-rag.lua